tests parameters refactor
