+++ tests/test_Net.py
@@ -1,10 +1,10 @@
 import torch
-import torch.nn as nn
 
 from model_constructor.net import Net, NewResBlock, ResBlock
 
-# from model_constructor.layers import SEModule, SimpleSelfAttention
+from .parameters import ids_fn
 
+# from model_constructor.layers import SEModule, SimpleSelfAttention
 
 bs_test = 4
 
@@ -22,19 +22,6 @@ params = dict(
     stem_bn_end=[True, False],
     stem_stride_on=[0, 1],
 )
-
-
-def value_name(value) -> str:  # pragma: no cover
-    name = getattr(value, "__name__", None)
-    if name is not None:
-        return name
-    if isinstance(value, nn.Module):
-        return value._get_name()  # pylint: disable=W0212
-    return value
-
-
-def ids_fn(key, value):
-    return [f"{key[:2]}_{value_name(v)}" for v in value]
 
 
 def pytest_generate_tests(metafunc):
