+++ tests/test_layers.py
@@ -1,16 +1,9 @@
 import torch
-import torch.nn as nn
 
-from model_constructor.layers import (
-    ConvBnAct,
-    Flatten,
-    Noop,
-    SEModule,
-    SEModuleConv,
-    SimpleSelfAttention,
-    noop,
-)
+from model_constructor.layers import (ConvBnAct, Flatten, Noop, SEModule,
+                                      SEModuleConv, SimpleSelfAttention, noop)
 
+from .parameters import ids_fn
 
 bs_test = 4
 
@@ -35,19 +28,6 @@ params = dict(
     rd_max=[False, True],
     use_bias=[True, False],
 )
-
-
-def value_name(value) -> str:
-    name = getattr(value, "__name__", None)
-    if name is not None:
-        return name
-    if isinstance(value, nn.Module):  # pragma: no cover
-        return value._get_name()  # pylint: disable=W0212
-    return value
-
-
-def ids_fn(key, value):
-    return [f"{key[:2]}_{value_name(v)}" for v in value]
 
 
 def pytest_generate_tests(metafunc):
