+++ tests/test_layers_depr.py
@@ -1,9 +1,9 @@
 # old (deprecated layers)
 import torch
-import torch.nn as nn
 
 from model_constructor.layers import ConvLayer, SEBlock, SEBlockConv
 
+from .parameters import ids_fn
 
 bs_test = 4
 
@@ -26,19 +26,6 @@ params = dict(
     bias=[False, True],
     groups=[1, 2],
 )
-
-
-def value_name(value) -> str:  # pragma: no cover
-    name = getattr(value, "__name__", None)
-    if name is not None:
-        return name
-    if isinstance(value, nn.Module):
-        return value._get_name()  # pylint: disable=W0212
-    return value
-
-
-def ids_fn(key, value):
-    return [f"{key[:2]}_{value_name(v)}" for v in value]
 
 
 def pytest_generate_tests(metafunc):
